third_party/WebKit/LayoutTests/external/wpt/css/css-tables-3/fixed-layout-excess-width-distribution-001-expected.txt

This is a testharness.js-based test.
FAIL #theTable 1 assert_equals: \n<table id="theTable">\n  <tbody><tr>\n    <td data-expected-width="180"></td>\n    <td data-expected-width="90"></td>\n    <td data-expected-width="30"></td>\n  </tr>\n</tbody></table>\nwidth expected 180 but got 100
Harness: the test ran to completion.
